rename(VJNP-99): 상수 관리 파일인 reaction을 feedCard로 변경
